src/: move code, that was intended to replace regex usage into `src/matching`
